requirements.txt
annotated-types==0.6.0
anyio==3.7.1
black==23.11.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
distro==1.8.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
mypy-extensions==1.0.0
openai==1.2.4
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
pydantic==2.5.0
pydantic_core==2.14.1
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
